1. application parameter injection replaced with context injection. 2. CoinApp moved to di package.
